# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AITeachMe",
    "description": "本地优先的 AI 助教后端服务。",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/subjects/{subject}/profile/mastery": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Mastery overview",
        "operationId": "api_get_mastery_overview_api_v1_subjects__subject__profile_mastery_get",
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MasteryOverviewResponse_"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects/{subject}/profile/mastery/unit/{teaching_unit_id}": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Unit mastery detail",
        "operationId": "api_get_unit_mastery_detail_api_v1_subjects__subject__profile_mastery_unit__teaching_unit_id__get",
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          },
          {
            "name": "teaching_unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Teaching Unit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MasteryStateResponse_"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects/{subject}/profile/mastery/node/{knowledge_node_id}": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Node mastery detail",
        "operationId": "api_get_node_mastery_detail_api_v1_subjects__subject__profile_mastery_node__knowledge_node_id__get",
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          },
          {
            "name": "knowledge_node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Knowledge Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MasteryStateResponse_"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects/{subject}/profile/review/tasks": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Pending review tasks",
        "operationId": "api_get_review_tasks_api_v1_subjects__subject__profile_review_tasks_get",
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_ReviewTaskResponse__"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects/{subject}/profile/review/tasks/{task_id}/complete": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Complete review task",
        "operationId": "api_complete_review_task_api_v1_subjects__subject__profile_review_tasks__task_id__complete_post",
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ReviewTaskResponse_"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiResponse_MasteryOverviewResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MasteryOverviewResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ApiResponse[MasteryOverviewResponse]",
        "example": {
          "code": 0,
          "data": {},
          "message": "ok"
        }
      },
      "ApiResponse_MasteryStateResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MasteryStateResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ApiResponse[MasteryStateResponse]",
        "example": {
          "code": 0,
          "data": {},
          "message": "ok"
        }
      },
...
      "ApiResponse_ReviewTaskResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewTaskResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ApiResponse[ReviewTaskResponse]",
        "example": {
          "code": 0,
          "data": {},
          "message": "ok"
        }
      },
...
      "ApiResponse_list_ReviewTaskResponse__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ReviewTaskResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ApiResponse[list[ReviewTaskResponse]]",
        "example": {
          "code": 0,
          "data": {},
          "message": "ok"
        }
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "type": "null",
            "title": "Data",
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ErrorResponse",
        "description": "统一错误响应结构。",
        "example": {
          "code": 404,
          "message": "资源不存在"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MasteryOverviewResponse": {
        "properties": {
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "weak_unit_count": {
            "type": "integer",
            "title": "Weak Unit Count"
          },
          "weak_node_count": {
            "type": "integer",
            "title": "Weak Node Count"
          },
          "unit_states": {
            "items": {
              "$ref": "#/components/schemas/MasteryStateResponse"
            },
            "type": "array",
            "title": "Unit States"
          },
          "node_states": {
            "items": {
              "$ref": "#/components/schemas/MasteryStateResponse"
            },
            "type": "array",
            "title": "Node States"
          }
        },
        "type": "object",
        "required": [
          "subject",
          "user_id",
          "weak_unit_count",
          "weak_node_count"
        ],
        "title": "MasteryOverviewResponse"
      },
      "MasteryStateResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "target_kind": {
            "type": "string",
            "title": "Target Kind"
          },
          "teaching_unit_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teaching Unit Id"
          },
          "knowledge_node_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Node Id"
          },
          "mastery_score": {
            "type": "number",
            "title": "Mastery Score"
          },
          "confidence_score": {
            "type": "number",
            "title": "Confidence Score"
          },
          "stability_score": {
            "type": "number",
            "title": "Stability Score"
          },
          "forgetting_due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forgetting Due At"
          },
          "review_priority": {
            "type": "number",
            "title": "Review Priority"
          },
          "total_attempts": {
            "type": "integer",
            "title": "Total Attempts"
          },
          "correct_attempts": {
            "type": "integer",
            "title": "Correct Attempts"
          },
          "last_attempt_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Attempt At"
          },
          "state_version": {
            "type": "integer",
            "title": "State Version"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "target_kind",
          "mastery_score",
          "confidence_score",
          "stability_score",
          "review_priority",
          "total_attempts",
          "correct_attempts",
          "state_version",
          "updated_at"
        ],
        "title": "MasteryStateResponse"
      },
...
      "ReviewTaskResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "target_kind": {
            "type": "string",
            "title": "Target Kind"
          },
          "teaching_unit_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teaching Unit Id"
          },
          "knowledge_node_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Node Id"
          },
          "priority": {
            "type": "number",
            "title": "Priority"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "interval_days": {
            "type": "integer",
            "title": "Interval Days"
          },
          "ease_factor": {
            "type": "number",
            "title": "Ease Factor"
          },
          "repetition_count": {
            "type": "integer",
            "title": "Repetition Count"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "source_exam_paper_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Exam Paper Id"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "subject",
          "target_kind",
          "priority",
          "status",
          "interval_days",
          "ease_factor",
          "repetition_count",
          "updated_at"
        ],
        "title": "ReviewTaskResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}